Fix old SMT theories: LRA, arith, datatypes, diff logic
- theory_lra: fix if/else-if for mutually exclusive int/real bound
  creation; fix assert_lower/upper stats counter condition; use
  thread_local for validate_eq reentrancy guard
- theory_arith_core: bounds-check random initial value index
- theory_datatype: add null check before theory_var lookup in
  mk_class_description to prevent crash on missing enode
- diff_logic: use thread_local for edge_id_data to avoid data races;
  add enable_edge return value check
- theory_dense_diff_logic: fix display_explanation signature; add
  null-safety for get_enode in internalize
- theory_diff_logic_def: fix del_atoms accessing wrong clause slot
  for ternary atoms; add fallthrough attributes
- theory_utvpi_def: fix edge weight for negated equality atoms

+++ src/smt/theory_dense_diff_logic.h
@@ -99,11 +99,11 @@ namespace smt {
         };
         
         struct cell_trail {
-            unsigned short m_source;
-            unsigned short m_target;
+            unsigned       m_source;
+            unsigned       m_target;
             edge_id        m_old_edge_id;
             numeral        m_old_distance;
-            cell_trail(unsigned short s, unsigned short t, edge_id old_edge_id, numeral const & old_distance):
+            cell_trail(unsigned s, unsigned t, edge_id old_edge_id, numeral const & old_distance):
                 m_source(s), m_target(t), m_old_edge_id(old_edge_id), m_old_distance(old_distance) {}
         };
